Refactor classes for PHP 8.2 and PSR
